requirements.txt
anyio==3.7.0
asyncpg==0.27.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
pydantic==1.10.8
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
